requirements.txt
mastodon.py==1.8.1
chromadb==0.4.15
google-generativeai==0.3.2
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
loguru==0.7.2
